pyproject.toml
[project]
name = "ralf"
version = "0.1.0"
description = "Retrieval-Augmented Layout Transformer for Content-Aware Layout Generation"
readme = "README.md"
requires-python = ">=3.9,<3.11"
authors = [{ name = "Dummy User", email = "dummy@gmail.com" }]

[project.optional-dependencies]
original-impl = [
    "datasets>=2.13.0,<3",
    "protobuf<=3.20.3",
    "gcsfs>=2023.1.0,<2024",
    "pillow>=9.5.0,<10",
    "setuptools>=68.0.0",
    "tensorboard>=2.12,<2.13",
    "omegaconf>=2.3.0,<3",
    "einops>=0.6.1,<0.7",
    "scipy<=1.10.1",
    "seaborn>=0.12.2,<0.13",
    "tensorflow>=2.12.0,<2.13",
    "tensorflow-datasets>=4.9.2,<5",
    "hydra-core>=1.3.2,<2",
    "opencv-python>=4.8.0.74,<5",
    "torch==1.13.1",
    "torchvision==0.14.1",
    "torch-tb-profiler>=0.4.1,<0.5",
    "faiss-cpu>=1.7.4,<2",
    "prdc>=0.2,<0.3",
    "pytorch-fid>=0.3.0,<0.4",
    "python-json-logger>=2.0.7,<3",
    "multiprocess>=0.70.12",
    "pyyaml>=6.0.1,<7",
    "timm>=0.9.5,<1",
    "rich>=13.5.2,<14",
    "gdown>=4.7.1,<5",
    "dreamsim>=0.1.3,<0.2",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-cov",
    "mypy",
    "ty",
    "pre-commit",
    "ipython<8.12.1",
    "ipykernel>=6.23.3,<7",
    "pytest-sugar>=1.1.1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

#[tool.uv]
#extra-index-url = ["https://download.pytorch.org/whl/cu117"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"ralf.train.config" = ["**/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=ralf --cov-report=term-missing --cov-fail-under=90"
markers = ["cuda: requires CUDA/GPU for execution"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E402", "E501", "E741", "F841"]
